requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
django-cprofile-middleware==1.0.5
django-crispy-forms==1.9.1
djangorestframework==3.11.0
idna==2.10
joblib==0.16.0
mysqlclient==2.0.0
numpy==1.19.0
pandas==1.0.5
pycodestyle==2.6.0
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.22.1
scipy==1.5.1
six==1.15.0
snakeviz==2.1.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
urllib3==1.25.9
wincertstore==0.2
xgboost==1.1.1
